+++ back/src/companies/resolvers/queries/companyInfos.ts
@@ -8,17 +8,19 @@ import { getInstallation } from "../../database";
 import { searchCompany } from "../../search";
 
 /**
- * This function is used to return public company
- * information for a specific siret or VAT number. It merge info
- * from Sirene or VIES vat database, S3ic database and TD without
- * exposing private TD info like securityCode, users, etc
+ * Recherche et renvoie les données diffusables
+ * sur une entreprise pour un numéro de SIRET ou de TVA
+ * Fusionnant les infos des bases Tracdéchets et S3IC
+ * si elles existent
+ * Renvoie le type CompanyPublic pour la query companyInfos
+ * et le type CompanySearchPrivate pour la query companyPrivateInfos
  *
  * @param siretOrVat
  */
 export async function getCompanyInfos(
   siretOrVat: string
 ): Promise<CompanyPublic | CompanySearchPrivate> {
-  if (siretOrVat === undefined || !siretOrVat.length) {
+  if (!siretOrVat) {
     throw new UserInputError(
       "Paramètre absent. Un numéro SIRET ou de TVA intracommunautaire valide est requis",
       {
@@ -35,10 +37,10 @@ export async function getCompanyInfos(
 }
 
 const companyInfosResolvers: QueryResolvers["companyInfos"] = async (
-  parent,
+  _,
   args
 ) => {
-  if (args.siret === undefined && args.clue === undefined) {
+  if (!args.siret && !args.clue) {
     throw new UserInputError(
       "Paramètre siret et clue absents. Un numéro SIRET ou de TVA intracommunautaire valide est requis",
       {